refactor(frontend): move splitLines into src/lib/
Place lines.ts under a `lib/` subdirectory.
