+++ internal/frontend/src/lib/lines.test.ts
