+++ internal/frontend/src/lib/lines.ts
